requirements.txt
aiogram>=3.0.0
pyserial==3.5
python-dotenv==1.0.0 